backend/requirements.txt
fastapi~=0.81.0
pandas~=1.5.3
requests~=2.31.0
python-multipart==0.0.7
uvicorn~=0.18.3
lightgbm==4.4.0
pyyaml~=6.0
scikit-learn~=1.3.1
optuna~=3.6.1
joblib~=1.3.2
numpy~=1.24.1
